Separate forecast locations from fetching
Re-use https connection for multiple requests
Remove an unneeded require
